Split chaotic tests
