+++ integration-tests/src/test/java/org/zowe/apiml/integration/ha/GatewayChaoticTest.java
@@ -14,7 +14,6 @@ import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.zowe.apiml.util.DiscoveryRequests;
 import org.zowe.apiml.util.categories.ChaoticHATest;
 import org.zowe.apiml.util.config.ConfigReader;
 import org.zowe.apiml.util.config.DiscoveryServiceConfiguration;
@@ -23,22 +22,20 @@ import org.zowe.apiml.util.config.GatewayServiceConfiguration;
 import org.zowe.apiml.util.http.HttpRequestUtils;
 
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.util.Collections;
 
 import static io.restassured.RestAssured.given;
 import static io.restassured.http.ContentType.JSON;
 import static org.apache.http.HttpStatus.SC_OK;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 import static org.zowe.apiml.util.SecurityUtils.getConfiguredSslConfig;
 
 /**
- * Verify behaviour of the application under chaotic testing
+ * Verify behaviour of the Gateway under chaotic testing
  */
 @ChaoticHATest
-public class ChaoticTest {
+public class GatewayChaoticTest {
 
     private GatewayServiceConfiguration gatewayServiceConfiguration;
     private DiscoveryServiceConfiguration discoveryServiceConfiguration;
@@ -47,10 +44,8 @@ public class ChaoticTest {
     private final String SHUTDOWN = "/application/shutdown";
     private int gatewayInstances;
     private int discoveryInstances;
-    private DiscoveryRequests discoveryRequests;
     private String username;
     private String password;
-    private String[] discoveryHosts;
     private String[] gatewayHosts;
 
     @BeforeEach
@@ -60,9 +55,7 @@ public class ChaoticTest {
         discoveryServiceConfiguration = environmentConfiguration.getDiscoveryServiceConfiguration();
         username = environmentConfiguration.getCredentials().getUser();
         password = environmentConfiguration.getCredentials().getPassword();
-        discoveryHosts = discoveryServiceConfiguration.getHost().split(",");
         gatewayHosts = gatewayServiceConfiguration.getHost().split(",");
-        discoveryRequests = new DiscoveryRequests(discoveryHosts[1]);
         gatewayInstances = gatewayServiceConfiguration.getInstances();
         discoveryInstances = discoveryServiceConfiguration.getInstances();
         RestAssured.config = RestAssured.config().sslConfig(getConfiguredSslConfig());
@@ -70,29 +63,6 @@ public class ChaoticTest {
 
     @Nested
     class GivenHASetUp {
-        @Nested
-        class whenOneDiscoveryServiceIsNotAvailable {
-            @Test
-            void serviceStillRegisteredToOtherDiscovery() throws URISyntaxException {
-                assumeTrue(gatewayInstances > 1 && discoveryInstances > 1);
-                shutDownDiscoveryInstance(discoveryHosts[0]);
-                assertThat(discoveryRequests.isApplicationRegistered("DISCOVERABLECLIENT"), is(true));
-            }
-
-            void shutDownDiscoveryInstance(String host) throws URISyntaxException {
-                //@formatter:off
-                given()
-                    .contentType(JSON)
-                    .auth().basic(username, password)
-                    .when()
-                    .post(HttpRequestUtils.getUriFromDiscovery(SHUTDOWN, host))
-                    .then()
-                    .statusCode(is(SC_OK))
-                    .extract().body().asString();
-                //@formatter:on
-            }
-        }
-
         @Nested
         class whenOneGatewayIsNotAvailable {
             @Test
